pyproject.toml
[tool.poetry]
name = "yasf"
version = "0.0.1"
authors = [
  "Mikael Giacomini <hello@mikaelgiacomini.com>",
]
description = "A set of tools to build vendor-agnostic Python serverless functions"
readme = "README.md"
license = "Apache License 2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache License 2.0",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.7"
boto3 = { version = "^1.25", optional = true }
google-cloud-firestore = { version = "^2.7", optional = true }

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^2.12"

[tool.poetry.extras]
aws = ["boto3"]
gcp = ["google-cloud-firestore", "flask"]

[tool.pytest.ini_options]
addopts = "--cov-report term-missing --cov=src/yasf"